+++ packages/node/src/behavior/system/controller/ControllerBehavior.ts
@@ -6,7 +6,7 @@
 
 import { Behavior } from "#behavior/Behavior.js";
 import { BasicInformationBehavior } from "#behaviors/basic-information";
-import { ConnectionlessTransportSet, ImplementationError, Logger } from "#general";
+import { ConnectionlessTransportSet, ImplementationError, Logger, SharedEnvironmentServices } from "#general";
 import { Node } from "#node/Node.js";
 import {
     Ble,
@@ -49,12 +49,15 @@ export class ControllerBehavior extends Behavior {
             throw new ImplementationError("adminFabricLabel must be set for ControllerBehavior");
         }
 
+        const node = Node.forEndpoint(this.endpoint);
+
         // Configure discovery transports
         if (this.state.ip === undefined) {
             this.state.ip = true;
         }
         if (this.state.ip !== false) {
-            this.env.get(ScannerSet).add((await this.env.load(MdnsService)).client);
+            this.internal.services = this.env.asDependent();
+            this.env.get(ScannerSet).add((await this.internal.services.load(MdnsService)).client);
         }
 
         if (this.state.ble === undefined) {
@@ -83,7 +86,6 @@ export class ControllerBehavior extends Behavior {
             }
         }
 
-        const node = Node.forEndpoint(this.endpoint);
         this.reactTo(node.lifecycle.online, this.#nodeOnline);
         if (node.lifecycle.isOnline) {
             this.#nodeOnline();
@@ -95,6 +97,7 @@ export class ControllerBehavior extends Behavior {
         await this.env.close(ActiveDiscoveries);
         this.env.delete(FabricAuthority);
         this.env.delete(ScannerSet);
+        await this.internal.services?.close();
     }
 
     get fabricAuthorityConfig(): FabricAuthorityConfiguration {
@@ -168,6 +171,8 @@ export namespace ControllerBehavior {
             commissionable: true,
             operationalTargets: [],
         };
+
+        services?: SharedEnvironmentServices;
     }
 
     export class State {