+++ packages/node/src/behaviors/general-diagnostics/GeneralDiagnosticsServer.ts
@@ -327,7 +327,8 @@ export class GeneralDiagnosticsServer extends Base {
     }
 
     async #updateNetworkList() {
-        const mdnsService = this.env.get(MdnsService);
+        await using services = this.env.asDependent();
+        const mdnsService = services.get(MdnsService);
         const mdnsLimitedToNetworkInterfaces = mdnsService.limitedToNetInterface;
 
         const networkRuntime = this.env.get(NetworkRuntime) as ServerNetworkRuntime;